database configuration + missing fields fix

+++ api/database/CRUD/delete.php
@@ -12,7 +12,8 @@ $database = new Database();
 $db = $database->getConnection();
   
 $results=array();
-if(isset($_POST['country'])&&isset($_POST['age'])&&isset($_POST['year'])&&isset($_POST['gender'])&&isset($_POST['BMI_type'])){
+if(isset($_POST['country'])&&isset($_POST['age'])&&isset($_POST['year'])&&isset($_POST['gender'])&&isset($_POST['BMI_type']) &&
+	$_POST['country'] != "" && $_POST['age'] != "" && $_POST['year'] != "" && $_POST['gender'] != "" && $_POST['BMI_type'] != ""){
 	$statement=$db->prepare("delete from data where BMI_type= ? and year = ? and country = ? and age = ? and gender = ?");
 	$statement->bind_param("sisss",$_POST['BMI_type'],$_POST['year'],$_POST['country'],$_POST['age'],$_POST['gender']);
